LayoutTests/fast/forms/switch/display-expected.txt

PASS getComputedStyle(control).display is "none"
PASS getComputedStyle(control).display is "inline-block"
PASS getComputedStyle(control).display is "block"
PASS getComputedStyle(control).display is "block"
PASS getComputedStyle(control).display is "inline-grid"
